pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent"
version = "0.1.0"
description = "A terminal-based AI coding agent for natural language code generation and execution"
authors = [{name = "AI Coding Agent", email = "agent@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "e2b>=0.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
    "gitpython>=3.1.0",
    "jedi>=0.19.0",
]

[project.scripts]
coding-agent = "coding_agent.cli:main"

[project.urls]
Homepage = "https://github.com/example/coding-agent"
Repository = "https://github.com/example/coding-agent.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["coding_agent*"]
